requirements.txt
tensorflow
pandas
gdown
dvc
mlflow
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .